backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
python-dotenv
chromadb
google-generativeai
python-docx
python-pptx
PyMuPDF
opencv-python
pytesseract